provisioning/adapter/incus: move to internal/adapter/incus
